requirements.txt
streamlit>=1.28.0
google-generativeai>=0.8.0
plotly>=5.18.0
reportlab>=4.0.0
kaleido>=0.2.1
python-dotenv>=1.0.0
